Added what-if scenarios
A SudokuWhatIf class was added that is used whenever the solver cannot figure out the complete sulutions and gets stuck. Users can then set a cell themselves to check whether there the solver can find a solution using this guess. It may be necessary, to enter several guesses, which however can lead to invalid configurations. This is just experimental for now.